requirements.txt
playwright==1.41.0
python-dotenv==1.0.0
schedule==1.2.1
playwright-stealth==1.0.6
